Example Character Select, Bug Fixes
Images are downloaded from the internet to the game now.

Fixed a bug in avatar_creator.js where you could submit a character
even if you didn’t have a player name or email. Changed it so that it
checks errors while you type.

Search will use the first letter as a starting db search and then use
the word algorithm to find best matches.

Added a library to convert JSON to lua objects and updated the credits
to give credit to the person.

Added some constant strings for the search URLs and stuff.